requirements.txt
openenv-core>=0.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
transformers>=4.45.0
torch>=2.4.0
jsonschema>=4.20.0
datasets>=2.20.0
huggingface_hub>=0.25.0
gradio==4.44.0
matplotlib>=3.8.0
pandas>=2.2.0
pytest>=8.0.0
